refactor: extract hfl scoring
